backend/requirements.txt
django>=5.0
djangorestframework
django-cors-headers
google-genai
python-dotenv
requests
beautifulsoup4
gunicorn
whitenoise
